Move command related code
Move the classes to a separate file, and move parsing logic to the Command class for better encapsulation
