+++ projects/07/vm_translator/parser.py
@@ -1,33 +1,6 @@
-import enum
-from typing import Iterator, Optional, Tuple, TextIO, Union
+from typing import Iterator, Tuple, TextIO
 
-
-class ArithmeticCommandClass(enum.Enum):
-    ADD = 'add'
-    SUB = 'sub'
-    NEG = 'neg'
-    EQ = 'eq'
-    GT = 'gt'
-    LT = 'lt'
-    AND = 'and'
-    OR = 'or'
-    NOT = 'not'
-
-
-class StackCommandClass(enum.Enum):
-    POP = 'pop'
-    PUSH = 'push'
-
-
-class Command:
-    def __init__(
-            self,
-            command_class: Union[ArithmeticCommandClass, StackCommandClass],
-            target_segment: Optional[str] = None,
-            index: Optional[int] = None):
-        self.command_class = command_class
-        self.target_segment = target_segment
-        self.index = index
+from command import Command
 
 
 class Parser:
@@ -54,15 +27,9 @@ class Parser:
 
     def _inspect_raw_command(self, raw_command: Tuple[str]) -> Command:
         number_of_elements = len(raw_command)
-        if number_of_elements == 1:
-            # Arithmetic
-            return Command(command_class=ArithmeticCommandClass(raw_command[0]))
-        elif number_of_elements == 3:
-            # Stack operation
-            return Command(
-                command_class=StackCommandClass(raw_command[0]),
-                target_segment=raw_command[1],
-                index=int(raw_command[2])
+        if number_of_elements not in {1, 3}:
+            raise ValueError(
+                f"Invalid command length ({number_of_elements}) for {raw_command}"
             )
-        else:
-            raise ValueError(f"Unrecognized command {raw_command}")
+
+        return Command(*raw_command)
